Resolve ambiguity with sysexData and sysexChannel
Add "universal" property to indicate whether sysex message is universal. These messages will contain a channel byte. Everything after that byte is considered sysexData - there's no change in the subdata that was getting returned prior to this change. In the even of a non-universal sysex message, channel will be 0, since channels aren't normally included in sysex messages, and sysexData will span a range starting with the byte following manufacturerID (prior to this change, that information was discarded).

That means: for universal messages subID's are included in sysexData, for non-universal messages, any device specific information (such as modelID, versionID or whatever manufactures decide to include) will be included in sysexData.